Template expansion does not really work
Processing as a page leads to an illegal array access.
If you set php to ignore errors, then the final output does not actually expand any templates.
This explains why this feature works in tests suite but not in the real world.
This same as previous pull.  Just re-based it for various self-serving reasons.