Fix parameter choices in tidy (#824)
publication-place: If any one of publication-place, place, or location is defined, the location will show after the title; if publication-place and place or location are defined, then place or location is shown before the title prefixed with "written at" and publication-place is shown after the title.

publication-date: Date of publication when different from the date the work was written. Displays only if year or date are defined and only if different, else publication-date is used and displayed as date.

orig-year: Original publication year; displays after the date or year. For clarity, please supply specifics. For example: |orig-year=First published 1859 or |orig-year=Composed 1904. Alias: origyear