Prevent the bot from mangling arXiV citations
TLDR, some of the recent issues like [1] and [2] are caused
by what appears to be a API change on ArXiV's end. The current
implementation assumes that the ArXiV URL data are ordered in
the order they were requested (in the $ids variable in line 270
in apiFunctions.php).

This patch adds a step that cleans up the ordering of the data
ArXiV returns so that before we start fixing templates, the $ids
array has the same ordering as the data returned by arXiV.

[1]: https://en.wikipedia.org/wiki/User_talk:Citation_bot#CitationBot_merging_info_from_two_completely_different_references
[2]: https://en.wikipedia.org/wiki/User_talk:Citation_bot#ArXiv_references_getting_completely_messed_up