Fix work=/title= when converting from web to book
Previously, when the bot converted a {{cite web}}
template to {{cite book}} template it would keep
the work=/title= parameters intact. However, this breaks
CS1 templates. This was brought up in [1] and [2].

This patch adds code to switch the work= parameter to the
title= parameter and copying the existing title parameter
into the chapter parameter as required by CS1.

[1]: https://en.wikipedia.org/wiki/User_talk:Citation_bot#Mangles_cite_web_to_cite_book_conversion_by_failing_to_change_parameter_names
[2]: https://en.wikipedia.org/wiki/User_talk:Citation_bot#Cite_book_to_cite_web_conversion_breaks_citation_by_failing_to_migrate_title=/work=_to_chapter=/title=